+++ Runtime/Scripts/ExpectedTypesOfResponse/FileResponse.cs
@@ -1,0 +1,12 @@
+﻿namespace KindMen.Uxios.ExpectedTypesOfResponse
+{
+    public sealed class FileResponse : ExpectedTypeOfResponse
+    {
+        public string Path;
+
+        public override void AddMetadataToRequest(Request request)
+        {
+            request.Headers.TryAdd("Accept", "application/octet-stream");
+        }
+    }
+}